backend/requirements.txt
Flask
python-dotenv
pandas # For handling CSV data later
sentence-transformers # For generating embeddings
# faiss-cpu==1.7.4 # For vector store (CPU version) # Ensure this or faiss-gpu is present
chromadb # Using ChromaDB for vector store
google-generativeai # For Gemini API
numpy # For numerical operations, especially with embeddings
# We will add more libraries like faiss-gpu if needed later
